requirements.txt
Flask
numpy
tensorflow
opencv-python
Pillow
requests
gunicorn
